Fixed promotion when playing against CPU
Promotion now works when playing against CPU. Fix was to defer the playCpuMove() function execution until the modal was closed. This additionally fixed the assumption in FEN string for CPU move generation, where it was assumed that the pawn was promoted to a queen. Now the server receives the correct FEN state and can make the best play according to the promotion